requirements.txt
typer[all]==0.12.3
Pillow>=10.3.0
opencv-python>=4.8.1
numpy>=1.26.0

# Optional (install manually for full local inference)
# diffusers==0.30.0
# transformers==4.43.3
# accelerate==0.33.0
# torch>=2.1.0
# safetensors>=0.4.2
# realesrgan
# basicsr
# timm>=0.9.2
